+++ CHANGELOG.md
@@ -14,4 +14,5 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip.
 
 ## [0.0.x](https://github.com/vsoch/cdb/master) (0.0.x)
+ - first release with basic data-container (0.0.1)
  - skeleton release (0.0.0)